feat: redisConfig 클래스 병합
