+++ src/main/java/com/example/twentyfiveframes/domain/RedisConfig.java
@@ -4,6 +4,7 @@ import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 
@@ -14,13 +15,29 @@ public class RedisConfig {
 
     @Bean
     public RedisConnectionFactory redisConnectionFactory() {
-        return new LettuceConnectionFactory(); // application.yml 설정 필요
+        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration("localhost", 6380);
+        config.setDatabase(0);
+        return new LettuceConnectionFactory(config); // application.yml 설정 필요
     }
 
     @Bean
     public RedisTemplate<String, Object> redisTemplate() {
         RedisTemplate<String, Object> template = new RedisTemplate<>();
         template.setConnectionFactory(redisConnectionFactory());
+        return template;
+    }
+
+    @Bean
+    public RedisConnectionFactory authRedisConnectionFactory() {
+        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration("localhost", 6380);
+        config.setDatabase(1);
+        return new LettuceConnectionFactory(config);
+    }
+
+    @Bean(name = "authRedisTemplate")
+    public RedisTemplate<String, String> authRedisTemplate() {
+        RedisTemplate<String, String> template = new RedisTemplate<>();
+        template.setConnectionFactory(authRedisConnectionFactory());
         return template;
     }
 }
