+++ src/main/java/com/example/twentyfiveframes/domain/user/dto/UserRequestDto.java
@@ -2,11 +2,13 @@ package com.example.twentyfiveframes.domain.user.dto;
 
 import com.example.twentyfiveframes.domain.user.entity.UserType;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 
 public class UserRequestDto {
 
     @Getter
+    @NoArgsConstructor(force = true)
     @RequiredArgsConstructor
     public static class Signup {
